strstr, strnstr, atoi

+++ ft_strchr.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-char	*ft_strchr(const char *s, int c)
+char		*ft_strchr(const char *s, int c)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
