+++ ft_strcmp.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-int		ft_strcmp(const char *s1, const char *s2)
+int			ft_strcmp(const char *s1, const char *s2)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] && s2[i])
